added values checking before deleting, updating and creating todo

+++ server/controllers/Todos/deleteTodo.js
@@ -2,6 +2,13 @@ import Todo from '../../models/todoModel'
 
 const deleteTodo = async (request, response, id) => {
   try {
+    const gotTodo = await Todo.findById(id)
+    if (!gotTodo) {
+      response.writeHead(404)
+      response.end(JSON.stringify('todo was not found'))
+      return
+    }
+
     const todo = await Todo.findByIdAndDelete(id)
     console.log('todo was deleted')
     response.writeHead(200)
